Time: 8 ms (58.21%), Space: 20.5 MB (45.25%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -12,13 +12,13 @@
 class Solution {
     public:
 
-    unordered_map<long, int> mp;
+    unordered_map<long long , int> mp;
     int count;
     int target;
-    void find(TreeNode* root, long sum) {
+    void find(TreeNode* root, long long sum) {
         if (root == 0) return;
         sum += root->val;
-        int prefixSum = sum - target;
+        long long prefixSum = sum - target;
         if (mp.count(prefixSum)) {
             count += mp[prefixSum];
         }
